generalize constraints
this patch is pretty hefty in that it changes the implementation of how
constraints were applied. Now all constraints follow a uniform model and
are implemented in terms of hooks:

- [shouldConsumeEvent] indicating whether the application event is
  relevant to the constraint
- [eventWasConsumed] giving a chance for the constraint to advance its
  state
- [shouldProducePoint] indicating whether a metric point should be
  submitted to the backend - that is, the constraint is fulfilled
- [didProducePoint] giving a chance for the constraint to further
  advance its state if necessary

the only drawback compared to the original implementation is that it's
hard to serialize the state now since we're keeping a hold of the
constraint hooks. However, it should be straightforward to implement a
serializer.

The API of the metric definition has changed a bit:

- "events" is no longer supported, in its place one should use the
  TypeConstraint interface
- "constraint" is no longer supported in lieu of the array "constraints"
  property
- "buffer" is no longer supported, in its place one should use any of
  the time/buffering constraints (such as RateConstraint or
  ParametricRateConstraint)